feat: initialize STS project with essential configurations and reset script
- Added package.json with dependencies and scripts for the STS project.
- Implemented reset-project.js script to reset the project structure and manage existing files.
- Created tsconfig.json for TypeScript configuration with strict mode and path aliases.
