+++ Frontend/STS/app.json
@@ -1,11 +1,11 @@
 {
   "expo": {
-    "name": "STS-App",
-    "slug": "STS-App",
+    "name": "STS",
+    "slug": "STS",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/images/icon.png",
-    "scheme": "stsapp",
+    "scheme": "sts",
     "userInterfaceStyle": "automatic",
     "newArchEnabled": true,
     "ios": {
@@ -19,7 +19,8 @@
         "monochromeImage": "./assets/images/android-icon-monochrome.png"
       },
       "edgeToEdgeEnabled": true,
-      "predictiveBackGestureEnabled": false
+      "predictiveBackGestureEnabled": false,
+      "package": "com.anonymous.STS"
     },
     "web": {
       "output": "static",
