+++ Frontend/STS/components/haptic-tab.tsx
