+++ Frontend/STS/components/ui/icon-symbol.ios.tsx
