+++ Frontend/STS/components/ui/icon-symbol.tsx
