+++ Frontend/STS/scripts/reset-project.js
